+++ packages/article/__tests__/web/semantic.web.test.js
@@ -95,6 +95,7 @@ const tests = [
               brightcovePolicyKey: "1.2.3.4",
               brightcoveAccountId: "57838016001",
               paidOnly: "false",
+              skySports: false,
               caption: "This is video caption",
               posterImageUrl: "https://image.io"
             },