+++ packages/card/card-wrapper.js
@@ -1,0 +1,2 @@
+export const CardWrapper = ({ children }) => children;
+export const ReversedCardWrapper = ({ children }) => children;