+++ packages/card/card.showcase.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { Text, View } from "react-native";
-import styled from "styled-components";
-import { breakpoints } from "@times-components/styleguide";
 import Card from "./src/card";
+import { CardWrapper, ReversedCardWrapper } from "./card-wrapper";
 
 const cardProps = {
   image: {
@@ -12,36 +11,6 @@ const cardProps = {
   imageRatio: 1.5,
   imageSize: 360
 };
-
-const CardWrapper = styled(View)`
-  @media (min-width: ${breakpoints.medium}px) {
-    .exampleCardImage {
-      flex-grow: 2 !important;
-      margin-bottom: 0;
-      min-width: 360px;
-      padding-right: 15px;
-    }
-    .exampleCardContent {
-      flex-grow: 2.7 !important;
-      min-width: 380px;
-    }
-  }
-`;
-
-const ReversedCardWrapper = styled(View)`
-  @media (min-width: ${breakpoints.medium}px) {
-    .exampleCardImage {
-      flex-grow: 2 !important;
-      margin-bottom: 0;
-      min-width: 360px;
-      padding-left: 15px;
-    }
-    .exampleCardContent {
-      flex-grow: 2.7 !important;
-      min-width: 380px;
-    }
-  }
-`;
 
 const childStyle = {
   borderColor: "black",