+++ packages/eslint-config-thetimes/index.js
@@ -45,6 +45,11 @@ module.exports = {
         object: "React",
         property: "Component",
         message: "Please destruct Component from React."
+      },
+      {
+        object: "React",
+        property: "PureComponent",
+        message: "Please destruct Component from React."
       }
     ]
   },