+++ packages/video/src/sky-sports-text-prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from "prop-types";
+
+const propTypes = {
+  text: PropTypes.string.isRequired
+};
+
+export default propTypes;