+++ packages/article-topics/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.2.20"></a>
+## [3.2.20](https://github.com/newsuk/times-components/compare/@times-components/article-topics@3.2.19...@times-components/article-topics@3.2.20) (2018-10-06)
+
+**Note:** Version bump only for package @times-components/article-topics
+
+
+
+
+
 <a name="3.2.19"></a>
 ## [3.2.19](https://github.com/newsuk/times-components/compare/@times-components/article-topics@3.2.18...@times-components/article-topics@3.2.19) (2018-10-03)