+++ packages/card/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.1.1"></a>
+## [4.1.1](https://github.com/newsuk/times-components/compare/@times-components/card@4.1.0...@times-components/card@4.1.1) (2018-10-06)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
 <a name="4.1.0"></a>
 # [4.1.0](https://github.com/newsuk/times-components/compare/@times-components/card@4.0.11...@times-components/card@4.1.0) (2018-10-03)