+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.5.47"></a>
+## [1.5.47](https://github.com/newsuk/times-components/compare/@times-components/pages@1.5.46...@times-components/pages@1.5.47) (2018-10-06)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 <a name="1.5.46"></a>
 ## [1.5.46](https://github.com/newsuk/times-components/compare/@times-components/pages@1.5.45...@times-components/pages@1.5.46) (2018-10-04)