+++ packages/provider-queries/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.4.0"></a>
+# [1.4.0](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@1.3.1...@times-components/provider-queries@1.4.0) (2018-10-06)
+
+
+### Features
+
+* add Sky sports banner ([#1382](https://github.com/newsuk/times-components/issues/1382)) ([231af44](https://github.com/newsuk/times-components/commit/231af44))
+
+
+
+
+
 <a name="1.3.1"></a>
 ## [1.3.1](https://github.com/newsuk/times-components/compare/@times-components/provider-queries@1.3.0...@times-components/provider-queries@1.3.1) (2018-09-27)