+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.41"></a>
+## [2.1.41](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.1.40...@times-components/video-label@2.1.41) (2018-10-06)
+
+**Note:** Version bump only for package @times-components/video-label
+
+
+
+
+
 <a name="2.1.40"></a>
 ## [2.1.40](https://github.com/newsuk/times-components/compare/@times-components/video-label@2.1.39...@times-components/video-label@2.1.40) (2018-10-03)